got rid of the old code and only kept the es6 portion
